pyproject.toml
[project]
name = "ner"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.14.20",
    "llama-index-graph-stores-neo4j>=0.7.0",
    "llama-index-llms-openai>=0.7.5",
    "neo4j>=5.28.3",
    "python-dotenv>=1.2.2",
    "fpdf",
    "pymupdf4llm>=1.27.2.2",
    "llama-index-embeddings-huggingface>=0.6.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.6.0,<2.7",
    "torchvision>=0.21.0,<0.22",
    "streamlit>=1.56.0",
    "arize-phoenix>=14.8.0",
    "openinference-instrumentation-llama-index>=4.3.9",
]

# CUDA 12.4 wheels (GPU). Without this, uv resolves torch from PyPI as +cpu.
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
